rename: apis-index.ts 파일 관심사별로 로직 분리 및 상수화된 엔드포인트 적용
