+++ src/apis/pickle.api.ts
@@ -1,0 +1,3 @@
+import client from '@/apis/axios';
+
+export const picklesRequests = Object.freeze({});
