+++ src/apis/types/auth.type.ts
